+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Added
 
+- Add API bridge connection backend for Open EdX Hawthorn
 - Add Linkedin badge in social networks templates.
 - Add new section "septenary" with a new wave decoration.
 - Add redis sentinel cache backend